Implement label support
All existing facilities do not allow to easily use dynamic string content,
so let's introduce a new concept - labels.

Labels are dynamic strings stored in a fixed-size ring buffer (256 Kb by
default). Strings are limited to 256 characters in size (configurable) to
simplify ring buffer implementation and to avoid buffer overfilling.

Each label has to be submitted with a group so that labels can be filtered
similar to timers.

For convenience a printf-style interface is also provided.