Add "Frame" display mode
This mode only shows the frame bar but does not show the scope bars. This
works well for when the profiler is not paused since the bars no longer
obscure the game and distract the user. If a frame spike happens it's
visible on the frame graph and at this point it's easy to pause and switch
to the detailed view.