Move web server to another thread
Now all socket operations are blocking and we only hold the mutex while
sending data via the socket.

This simplifies the flow and also fixes a case when a browser (e.g.
Chrome) connects to the socket that we open and does not write data to it
for 10+ seconds, which causes severe stalls on the main thread.