Streamline GPU implementation
The most common case is not using GPU timers or using built-in
implementations; we no longer require anything for the first case, and
there's just one define for GL that automatically picks one of the
implementations.

For custom implementations a new define MICROPROFILE_GPU_TIMERS_CUSTOM has
to be defined, in which case the application has to implement all 6
functions.