Synchronize GPU and CPU timers when using TIME_ELAPSED
Since TIME_ELAPSED does not track GPU idle, in CPU-bound cases GPU frames
are queued right after each other without any gaps which is not what actually
happens...

To resolve this, insert a fake timer on GpuFlip that has the current CPU tick
converted to GPU timer frequency. This makes sure that all GPU frames start on
GpuFlip which is a reasonable approximation.