Simplify D3D11 timer implementation
Use the same approach as GL implementation - read data from the query when
the caller requests it instead of on flip. This allows us to use the
generic frame logic instead of implementing 3-frame delay again.

Additionally, we don't need to read frequency every frame - since we don't
use Disjoint flag currently (and it's unclear how to do that efficiently)
and store timers in ticks, frequency changes will result in some
disruption in data anyway.

So we still try to read it every frame but only if the data has arrived,
which lets us use just one query.