Fix pause mode
Commit e6d6c0e9f0c473a9346b7d859feb7d21ceb06c84 changed pause mode
behavior to record data; this is bad since it means that in pause mode you
start to lose labels and region data as the buffers are being written to.

It's more correct to preserve the original logic *except* for nDisplay
check which makes the profiler run if you enable the groups even if it's
not rendering (which is what we want!).