Setup socket to avoid SIGPIPE
Unfortunately, there are two different ways to do this - BSD way (that works
on OSX/iOS), which conveniently requires a one-time socket setup, and Linux
way, that requires passing custom flags to every single send() call.

To make this work we move send() call to MicroProfileSendSocket that uses the
correct flags.