core: Fix MicroProfileFlip reentrancy
MicroProfileFlip initializes the web server which on some systems can execute
random callbacks dispatched to the thread (e.g. APC). If one of those is the
function that calls MicroProfileFlip we can initialize the web server several
times which does not end well.

This is a very weird edge case but it's trivial to fix.