core: Immediately activate new groups as necessary
When you use a token for the first time, the bit in the group mask is set.

However, when all groups are active and profiler is running, this bit will
only propagate to the active mask on next flip, causing the first frame of
data for new groups to be ignored.

To fix this we immediately set the bit under the same conditions that flip
sets it in.