core: Lazily allocate thread log buffers
This makes adding extra threads and naming them very cheap as long as you're
not logging events. Additionally, it means that profiler consumes less extra
memory if it's compiled in *but* all groups are disabled.

In theory this makes it possible to have per-thread buffer sizes which is very
interesting for some use cases but requires extra work.