ui: Refactor CPU-GPU sync
Store the reference ticks and only request new data if we need to.

We update if the reference tick is before the latest frame in the snapshot
(since we're paused nFrameStartCpu does not increase). This maximizes the
precision of the last frame; we will not get better data by requesting other
sync points in the paused view anyway.

This behavior is consistent with web view as well - the most recent frame is
the one that has highest quality data.