core: Implement external ETW context switch traces
This makes the cswitch implementation similar to that for OSX -
microprofile itself does not have tracing code; you need to launch an
external application that requires elevation, and it communicates the
tracing results via a named pipe.

The application can run indefinitely - it recreates the pipe when the
original pipe is broken.

Unfortunately ETW data does not have PID data so there's some complexity
associated with caching TID->PID lookup.