core: Refactor GL timer implementation
Since we have a fixed small number of GPU frames it does not make as much sense
to maintain a ring buffer - it's easier to just split the query range between
frames. This makes it easier to handle overflows in multithreaded submission
case and hopefully will make it possible to write code that looks similar
between different APIs (obviously modulo synchronization issues and the like).