core: Implement GetGpuTickReference for GL/OSX
The UI/web logic in absence of GPU synchronization assumes that the GPU work
for frame N starts executing at the beginning of frame N, which obviously can't
happen.

GL implementation is effectively faking synchronization at every flip anyway so
it makes sense to pretend we have GPU sync working - this makes it so that the
GPU work starts executing immediately after flip.