core: Fix GPU timer queue management
The constant MICROPROFILE_GPU_FRAME_DELAY specifies the latency in getting
GPU data. If it's set to 0 this means that we're reading the results that
we have just submitted.

To make the frame management work, we need to introduce a derivative
constant, MICROPROFILE_GPU_FRAMES, that is delay+1 and is used to segment
the available query space and synchronize in the implementation.

Finally, Flip was mismanaging the frames, reading the start frame marker
from one frame ahead of the actual GPU data. Change the code to read all
data from a single GPU frame in one go.