ui: Fix drawing early frames being expensive
The drawing code basically rendered all log entries from the selected frame up
until the current frame. This manifests in two places:

1. The range of frames/log entries being selected ends at current frame
2. We don't do any early outs during rendering

Point 1 is problematic and will be fixed later; however, there was a solution
for point 2 which was reverted in 92ce16c33690a555a75bd68c634a30e14d72370e.

As far as I can tell the only issue that can occur on GPU is early out because
the GPU submission order does not match the order of calls to GpuSubmit which
is a user error.