ui: Select correctly bounded frame range for rendering
Instead of just scanning backwards up until we have the first visible frame,
bracket the frame range using the end ticks as well, and only render frames
in a given range.

This is somewhat redundant given the early-out, but since early-out may have
side-effects that we'll have to counter later this is still useful.

Note that scanning region is extended by "gap time" (50 ms). On the left hand
side the gap time is useful so that we can build up the stack of events properly
in case some events span the frame boundary. On the right hand side the gap
time is useful so that we can close events that span the frame boundary.

The early-out thus is still a bit useful to reduce the size of right-hand side
gap, although it's not super critical - maybe we'll remove it later...

Note that when scanning backwards we still need to discard frames that have
events that have been overwritten by later data. It's cumbersome to do that
due to ring buffer structure and us discarding the high bits from ring buffer
indices.