ui: Use GPU frame time for GPU frame selection
Since GPU frames are generally offset from CPU frames, when gap time is 0 the
frame selection logic does not pick the GPU frame with the proper regions if
only one frame is visible.

Gap time sometimes hides this, but in some cases the GPU delay can be big enough
so that the issue is still visible.

Fix this by using GPU or CPU time range for frame range logic.