Use dynamic dispatch for GPU timer API
In some applications it's important to be able to select rendering API
implementations dynamically, for example to support DX11 and DX12 in the
same executable. Currently this is impossible because GPU API is
selected using exclusive compile time defines.

Rework the timer implementations so that there's just one active at a
single time, but you can select multiple implementations to be compiled
in by defining several macros. All functions are routed through function
pointers in the g_MicroProfile.GPU table, with appropriate defaults for
implementations that choose to not implement e.g. CPU-GPU sync.

This also makes it a bit more straightforward to integrate custom timers -
you just fill the function table with your implementation, but you're free
to track as much external global state as you wish since *all* state
you'll need is global anyway. Thus you don't need to define anything to
use custom timers.

There's a tiny bit of extra overhead due to indirect dispatch but it's
insignificant compared to the actual work that the driver would do for
most of the calls.