Fix some MICROPROFILE_ENABLED=0 macros
There were some inconsistencies with availability and/or set of
parameters for some macros when microprofile was disabled; this change
fixes the obvious ones.

The GpuInit*/Shutdown are still not macroed out; the rationale is that
it's easy to surround the requisite init code with macros on the user
side, but this might change in the future.