Fix crash in UI code when a thread exits
We are no longer preserving thread log data when thread exits to
conserve memory; this makes pointers to thread logs/thread log entries
unsafe to keep over multiple frames.

UI code used one persistent pointer (to be able to do just one pass over
the data to draw bars) into the thread log; the use of this pointer was
relatively safe as it's only used for equality comparison inside the
loop with another valid log entry, but we did need to read the timer
index to determine if other entries we process need to be colored
differently. This change caches the timer index in a variable to avoid
use-after-free here.