Add support for label literals
Labels are convenient to store arbitrarily formatted strings, however
they have two issues:

- They require copying the string data to profiler storage
- They use a fixed size buffer; overflowing this buffer causes labels
submitted earlier to disappear

In some cases, we need to attach metadata in the form of literal
strings - that is, strings with lifetime that surpasses that of
profiling frames. For example, annotating scripting language bindings
generally means that there are too many unique strings to create a timer
for each, but these strings are never destroyed.

In these cases it's now possible to use label literals - the pointer to
the string will be stored in the thread log directly (using 48 bits for
storage), and can be accessed via the same UI that works for regular
labels.

Currently the HTML dump doesn't deduplicate label literals, so the dump
will still contain all label copies - something we can remedy later.